+++ packages/api-headless-cms/package.json
@@ -27,6 +27,7 @@
     "@graphql-tools/schema": "^10.0.34",
     "@webiny/api": "0.0.0",
     "@webiny/api-core": "0.0.0",
+    "@webiny/api-testing": "0.0.0",
     "@webiny/di": "^1.0.2",
     "@webiny/error": "0.0.0",
     "@webiny/event-handler-core": "0.0.0",
