+++ packages/api-headless-cms/src/testing/createCmsTestHandler.ts
@@ -17,10 +17,10 @@ import type { SecurityPermission } from "@webiny/api-core/types/security.js";
 import type { DecryptedWcpProjectLicense } from "@webiny/wcp/types.js";
 import { HeadlessCmsFeature } from "~/HeadlessCmsFeature.js";
 import type { HeadlessCmsStorageOperations } from "~/types/types.js";
-import { TestIdentity, TestAuthenticator } from "./mocks/TestAuthenticator.js";
-import { TestPermissions, TestAuthorizer } from "./mocks/TestAuthorizer.js";
-import { AuthTriggerHandler } from "./handlers/AuthTriggerHandler.js";
-import { RootTenantInitializer } from "./handlers/RootTenantInitializer.js";
+import { TestIdentity, TestAuthenticator } from "@webiny/api-testing";
+import { TestPermissions, TestAuthorizer } from "@webiny/api-testing";
+import { AuthTriggerHandler } from "@webiny/api-testing";
+import { RootTenantInitializer } from "@webiny/api-testing";
 import { processLegacyPlugins } from "./processLegacyPlugins.js";
 
 const DEFAULT_IDENTITY: IdentityData = {
