+++ packages/api-headless-cms/src/testing/index.ts
@@ -1,7 +1,7 @@
 export { createCmsTestHandler } from "./createCmsTestHandler.js";
 export type { CmsTestHandlerParams, InvokeParams } from "./createCmsTestHandler.js";
 export { processLegacyPlugins } from "./processLegacyPlugins.js";
-export { TestIdentity, TestAuthenticator } from "./mocks/TestAuthenticator.js";
-export { TestPermissions, TestAuthorizer } from "./mocks/TestAuthorizer.js";
-export { AuthTriggerHandler } from "./handlers/AuthTriggerHandler.js";
-export { RootTenantInitializer } from "./handlers/RootTenantInitializer.js";
+export { TestIdentity, TestAuthenticator } from "@webiny/api-testing";
+export { TestPermissions, TestAuthorizer } from "@webiny/api-testing";
+export { AuthTriggerHandler } from "@webiny/api-testing";
+export { RootTenantInitializer } from "@webiny/api-testing";
