+++ packages/api-headless-cms/tsconfig.json
@@ -4,6 +4,7 @@
   "references": [
     { "path": "../api" },
     { "path": "../api-core" },
+    { "path": "../api-testing" },
     { "path": "../error" },
     { "path": "../event-handler-core" },
     { "path": "../feature" },
@@ -29,6 +30,8 @@
       "@webiny/api": ["../api/src"],
       "@webiny/api-core/*": ["../api-core/src/*"],
       "@webiny/api-core": ["../api-core/src"],
+      "@webiny/api-testing/*": ["../api-testing/src/*"],
+      "@webiny/api-testing": ["../api-testing/src"],
       "@webiny/error/*": ["../error/src/*"],
       "@webiny/error": ["../error/src"],
       "@webiny/event-handler-core/*": ["../event-handler-core/src/*"],
