+++ packages/api-testing/src/index.ts
@@ -1,5 +1,6 @@
 export { TestIdentity, TestAuthenticator } from "./mocks/TestAuthenticator.js";
 export { TestPermissions, TestAuthorizer } from "./mocks/TestAuthorizer.js";
+export type { TestPermissionsHolder } from "./mocks/TestAuthorizer.js";
 export { AuthTriggerHandler } from "./handlers/AuthTriggerHandler.js";
 export { RootTenantInitializer } from "./handlers/RootTenantInitializer.js";
 export { createIdentity } from "./identity.js";
